refactor: share registration types and ServerInfo between dwn-clients and dwn-server
Move ProofOfWorkChallengeModel, RegistrationData, and RegistrationRequest
from @enbox/dwn-server into @enbox/dwn-clients as shared types. Both the
client (DwnRegistrar) and server now import from the same source of truth.

Expand ServerInfo to include all fields returned by the /info endpoint:
url, server, version, sdkVersion (previously discarded by the client).
Update dwn-server's #handleInfo() to use the shared ServerInfo type.

Type DwnRegistrar's inline object literals with the shared types and
remove the TODO referencing issue #162.

Add @enbox/dwn-clients README and add the package to the monorepo README
coverage table and dependency graph.
