+++ packages/dwn-clients/src/index.ts
@@ -4,6 +4,7 @@ export * from './dwn-server-info-cache-memory.js';
 export * from './http-dwn-rpc-client.js';
 export * from './json-rpc.js';
 export * from './json-rpc-socket.js';
+export * from './registration-types.js';
 export * from './rpc-client.js';
 export * from './server-info-types.js';
 export * from './utils.js';
