--- packages/dwn-server/src/registration/proof-of-work-types.ts
@@ -1,7 +1,0 @@
-/**
- * Proof-of-work challenge model returned by the /registration/proof-of-work API.
- */
-export type ProofOfWorkChallengeModel = {
-  challengeNonce: string;
-  maximumAllowedHashValue: string;
-};
